2016/i365-td_server-013d94e67b8d/v1/requirements.txt
tushare
html5lib
bs4
pandas
lxml
eve
flask_oauthlib
redis
bcrypt
cattle
Flask-Bootstrap
eve-docs
pyOpenSSL
deco
Flask-Bootstrap